+++ test/fixtures/generator/bwx_spec.json
@@ -363,6 +363,7 @@
                     "tools/generator/test/XCScheme+ExtensionsTests.swift",
                     "tools/generator/test/XCSchemeEnvironmentVariablesExtensionsTests.swift",
                     "tools/generator/test/XCSchemeInfo+BuildActionInfoTests.swift",
+                    "tools/generator/test/XCSchemeInfo+BuildTargetTests.swift",
                     "tools/generator/test/XCSchemeInfo+HostInfoTests.swift",
                     "tools/generator/test/XCSchemeInfo+LaunchActionInfoTests.swift",
                     "tools/generator/test/XCSchemeInfo+ProfileActionInfoTests.swift",
