+++ tools/generator/src/Generator/XCSchemeInfo+TargetInfo.swift
@@ -173,38 +173,6 @@ extension Sequence where Element == XCSchemeInfo.TargetInfo {
     }
 }
 
-// extension Sequence where Element == XCSchemeInfo.BuildTarget {
-//     /// Return all of the buildable references for all of the target infos.
-//     var buildableReferences: [XCScheme.BuildableReference] {
-//         return flatMap(\.buildableReferences).inStableOrder
-//     }
-// }
-
-// extension Sequence where Element == XCSchemeInfo.BuildTarget {
-//     /// Return all of the `BuildAction.Entry` values.
-//     var buildActionEntries: [XCScheme.BuildAction.Entry] {
-//         // GH573: Pass the buildFor information to BuildAction.Entry init
-//         return buildableReferences.map { .init(withDefaults: $0) }
-//     }
-// }
-
-// TODO(chuck): Move this to BuildTarget file
-extension Sequence where Element == XCSchemeInfo.BuildTarget {
-    /// Return all of the `BuildAction.Entry` values.
-    var buildActionEntries: [XCScheme.BuildAction.Entry] {
-        // Create a (BuildableReference, BuildFor) for all buildable references
-        return flatMap { buildTarget in
-                return buildTarget.targetInfo.buildableReferences.map { ($0, buildTarget.buildFor) }
-            }
-            // Sort by in stable order by blueprint name
-            .sortedLocalizedStandard(\.0.blueprintName)
-            // GH573: Pass the buildFor information to BuildAction.Entry init
-            // TODO(chuck): Do I need to dedupe and merge buildFor due to the same host appearing
-            // multiple times in the list?
-            .map { buildableReference, _ in .init(withDefaults: buildableReference) }
-    }
-}
-
 extension Sequence where Element == XCSchemeInfo.TargetInfo {
     func buildPreActions(buildMode: BuildMode) throws -> [XCScheme.ExecutionAction] {
         let targetInfos = inStableOrder
